use bf16

+++ infer_engine.py
@@ -160,7 +160,7 @@ class TransfomersEngine(DeployEngine):
             load_in_8bit = False if config.load_in_4bit else config.load_in_8bit,
             llm_int8_threshold = 6.0,
             llm_int8_has_fp16_weight = False,
-            bnb_4bit_compute_dtype = torch.float16,     # 4位精度计算的数据类型。这里设置为torch.float16，表示使用半精度浮点数。
+            bnb_4bit_compute_dtype = torch.bfloat16,    # 4位精度计算的数据类型。这里设置为torch.bfloat16，表示使用半精度浮点数。
             bnb_4bit_quant_type = 'nf4',                # 4位精度量化的类型。这里设置为"nf4"，表示使用nf4量化类型。 nf4: 4bit-NormalFloat
             bnb_4bit_use_double_quant = True,           # 是否使用双精度量化。如果设置为True，则使用双精度量化。
         )
@@ -168,7 +168,7 @@ class TransfomersEngine(DeployEngine):
         # 创建模型
         self.model = AutoModelForCausalLM.from_pretrained(
             config.pretrained_model_name_or_path,
-            torch_dtype = torch.float16,
+            torch_dtype = torch.bfloat16,
             trust_remote_code = True,
             device_map = 'auto',
             low_cpu_mem_usage = True,   # 是否使用低CPU内存,使用 device_map 参数必须为 True
